More tests for the A64 entrypoints
